Finished version 1. Works as intended, with some bugs.
CTA notation must be in the following listed format:

Cta Q = Init q0;q0 z!a(x <= 2) q1;
Cta Z = Init z0;z0 q?a(x <= 2) z1;

Imperative details are:
'Cta', 'Init' and all of the spacing must be exactly the same.
The time condition must be in the same format that Go uses for equality.